requirements.txt
# --- Core Deep Learning ---
torch>=2.3.0
torchvision>=0.18.0
torchaudio>=2.3.0
transformers==4.45.0
accelerate==0.33.0
trl==0.10.1
peft==0.11.1
bitsandbytes==0.43.1
deepspeed==0.15.4
sentencepiece
tokenizers>=0.19.0
safetensors>=0.6.2

# --- Math / Science stack ---
numpy>=1.26.4
scipy>=1.12.0
sympy
einops
lion-pytorch
matplotlib
tqdm
packaging
typing_extensions
protobuf

# --- Data / Dataset ---
datasets>=3.0.0
pandas
sqlite-utils
h5py

# --- NLP / Geometry / Custom TGDK modules ---
# (These are local .py files in your repo; kept for clarity)
# Mahadevi.py
# Maharaga.py
# Trinity.py
# Duo_o1.py
# device.py
# volumetric_infinitizer.py
# possessor_module.py
# code_wright.py
# mirrorblade.py
# symbolic_adapter.py
# tgdk_datsik.py
# edge_of_fold.py
# icu.py
# mainmethods.py
# csr.py
# tvr.py
# rsicommand.py

# --- Cryptography / Security ---
cryptography>=42.0.0
pycryptodome
hashlib
psutil

# --- File / Document handling ---
pypdf
lzma
base58
pathlib
glob2
shutilwhich

# --- Web / API (optional, used by OliviaAI/TGDK bridges) ---
fastapi
uvicorn
requests

# --- Visualization / Control ---
tensorboard
tensorboardX
matplotlib-inline

# --- Optional system integrations ---
ninja
cmake
build
setuptools>=68.0.0
wheel>=0.43.0

# --- Database / IO ---
psycopg2-binary
SQLAlchemy>=2.0.0

# --- Logging / Diagnostics ---
loguru
rich
colorlog
